+++ c/jay/open_jay.cc
@@ -91,7 +91,7 @@ static void initStats(Stats* stats, const jay::Column* jcol) {
   auto tstats = static_cast<NumericalStats<T, A>*>(stats);
   auto jstats = static_cast<const JStats*>(jcol->stats());
   if (jstats) {
-    tstats->set_countna(static_cast<int64_t>(jcol->nullcount()));
+    tstats->set_countna(jcol->nullcount());
     tstats->set_min(jstats->min());
     tstats->set_max(jstats->max());
   }
